Add files via upload
How was prepared war in the Roman Empire? This text adventure game wants to dive into this question and explore a piece of history of one the greatest empires ever existed.
The game is elaborated on Python and divided in five steps. It comprehends multiple choices, randomization and a very hard final step to complete the adventure.