Add files via upload
A text adventure elaborated in Python to introduce to the preparation of war in one of the greatest empires ever existed: Rome.
The player will confront five steps to prepare a mission and successfully complete the game. To successfully achieve the end there will be random elements and strategic choices to take.